requirements.txt
selenium>=4.15.0
webdriver-manager>=4.0.1
pandas>=2.1.0
requests>=2.31.0
beautifulsoup4>=4.12.0
openai>=1.6.0
python-dotenv>=1.0.0
gspread>=6.0.0
google-auth>=2.25.0
